Move and Refactor convert to Template (#2349)
* Move convert from Site/index.ts to template.ts

* Remove redundant code

* Move readSiteConfig to SiteConfig.ts

* Move lodash const to constants.ts

* Update readSiteConfig comments

* Rename template to templatePath

* Replace then/catch with try/catch
