+++ packages/cli/src/cmd/init.js
@@ -2,11 +2,10 @@ const fs = require('fs-extra');
 const path = require('path');
 
 const { Template } = require('@markbind/core');
-const { Site } = require('@markbind/core').Site;
 
 const logger = require('../util/logger');
 
-function init(root, options) {
+async function init(root, options) {
   const rootFolder = path.resolve(root || process.cwd());
 
   if (options.convert) {
@@ -17,28 +16,25 @@ function init(root, options) {
   }
 
   const template = new Template(rootFolder, options.template);
-  template.init()
-    .then(() => {
-      logger.info('Initialization success.');
-    })
-    .then(() => {
-      if (options.convert) {
-        logger.info('Converting to MarkBind website.');
-        const outputRoot = path.join(rootFolder, '_site');
-        new Site(rootFolder, outputRoot).convert()
-          .then(() => {
-            logger.info('Conversion success.');
-          })
-          .catch((error) => {
-            logger.error(error.message);
-            process.exitCode = 1;
-          });
-      }
-    })
-    .catch((error) => {
-      logger.error(`Failed to initialize site with given template with error: ${error.message}`);
+
+  try {
+    await template.init();
+    logger.info('Initialization success.');
+  } catch (error) {
+    logger.error(`Failed to initialize site with given template with error: ${error.message}`);
+    process.exitCode = 1;
+  }
+
+  if (options.convert) {
+    logger.info('Converting to MarkBind website.');
+    try {
+      await template.convert();
+      logger.info('Conversion success.');
+    } catch (error) {
+      logger.error(error.message);
       process.exitCode = 1;
-    });
+    }
+  }
 }
 
 module.exports = {
